requirements.txt
allure-pytest==2.8.21
allure-python-commons==2.8.21
apipkg==1.5
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
execnet==1.8.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
paramiko==2.7.2
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.2
pytest-forked==1.3.0
pytest-xdist==2.1.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.3
